+++ OpenIddictExternalAuthentication/Cookies/RefreshTokenFromCookiesHandler.cs
@@ -1,0 +1,41 @@
+﻿using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
+using OpenIddict.Abstractions;
+using OpenIddict.Server;
+
+namespace Shaddix.OpenIddict.ExternalAuthentication.Cookies;
+
+/// <summary>
+/// Copies RefreshToken from Cookies to OpenId Request object
+/// </summary>
+public class RefreshTokenFromCookiesHandler
+    : IOpenIddictServerHandler<OpenIddictServerEvents.ValidateTokenRequestContext>
+{
+    private readonly IHttpContextAccessor _httpContextAccessor;
+
+    public RefreshTokenFromCookiesHandler(IHttpContextAccessor httpContextAccessor)
+    {
+        _httpContextAccessor = httpContextAccessor;
+    }
+
+    public ValueTask HandleAsync(OpenIddictServerEvents.ValidateTokenRequestContext context)
+    {
+        // Check if the request is for token exchange and contains a refresh token grant type
+        if (context.Request?.IsRefreshTokenGrantType() == true)
+        {
+            if (string.IsNullOrEmpty(context.Request.RefreshToken))
+                // Read the refresh token from the cookie
+                if (
+                    _httpContextAccessor.HttpContext.Request.Cookies.TryGetValue(
+                        StoreAccessRefreshTokenInCookieHandler.RefreshTokenCookieName,
+                        out var refreshToken
+                    )
+                )
+                {
+                    context.Request.RefreshToken = refreshToken;
+                }
+        }
+
+        return ValueTask.CompletedTask;
+    }
+}